fix(qa): repair live selector CI
